+++ src/app/layout.tsx
@@ -1,0 +1,10 @@
+import "@/styles/globals.css";
+import { ReactNode } from "react";
+
+export default function RootLayout({ children }: { children: ReactNode }) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  );
+}
