+++ src/components/TmpContext.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { ReactNode, createContext, useContext, useState } from "react";
 
 // jotai 대체를 위한 임시 context
