--- src/pages/_app.tsx
@@ -1,13 +1,0 @@
-import "@/styles/globals.css";
-import type { AppProps } from "next/app";
-import { QueryClient, QueryClientProvider } from "react-query";
-
-const queryClient = new QueryClient();
-
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <QueryClientProvider client={queryClient}>
-      <Component {...pageProps} />;
-    </QueryClientProvider>
-  );
-}
